# public/swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "University of Oslo Library Services Middleware (LSM)",
        "version": "0.2.0"
    },
    "paths": {
...
        "/subjects/search": {
            "get": {
                "tags": [
                    "Authorities"
                ],
                "summary": "Search authority records using the Skosmos API",
                "description": "Search for terms, optionally filtered by vocabulary and concept type.",
                "operationId": "App\\Http\\Controllers\\SubjectsController::search",
                "parameters": [
                    {
                        "name": "query",
                        "in": "query",
                        "description": "Case-insensitive search term. Use * at the beginning and/or end to truncate.",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "parent",
                        "in": "query",
                        "description": "Only search children of this concept, specified by URI.",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "group",
                        "in": "query",
                        "description": "Only search children of this group, specified by URI.",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "fields",
                        "in": "query",
                        "description": "Space-separated list of extra fields to include in the results. Supported values: 'broader'",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "unique",
                        "in": "query",
                        "description": "Boolean flag to indicate that each concept should be returned only once, instead of returning all the different ways it could match (for example both via prefLabel and altLabel).",
                        "schema": {
                            "type": "boolean",
                            "default": false
                        }
                    },
                    {
                        "name": "lang",
                        "in": "query",
                        "description": "Search language.",
                        "schema": {
                            "type": "string",
                            "enum": [
                                "nb",
                                "nn",
                                "en"
                            ]
                        }
                    },
                    {
                        "name": "labellang",
                        "in": "query",
                        "description": "Language used to format results.",
                        "schema": {
                            "type": "string",
                            "enum": [
                                "nb",
                                "nn",
                                "en"
                            ]
                        }
                    },
                    {
                        "name": "vocab",
                        "in": "query",
                        "description": "Subject vocabulary. Leave blank to search all subject vocabularies.",
                        "schema": {
                            "type": "string",
                            "enum": [
                                "realfagstermer",
                                "humord",
                                "tekord",
                                "mrtermer",
                                "usvd",
                                "lskjema",
                                "ddc"
                            ]
                        }
                    },
                    {
                        "name": "type",
                        "in": "query",
                        "description": "All resources have type `Concept`or `Facet`. Concepts are further subdivided into `Topic`, `Place`, `Time`, `CompoundConcept`, `VirtualCompoundConcept` and `NonIndexable`.",
                        "schema": {
                            "type": "string",
                            "enum": [
                                "Concept",
                                "Facet",
                                "Topic",
                                "Place",
                                "Time",
                                "CompoundConcept",
                                "VirtualCompoundConcept",
                                "NonIndexable"
                            ]
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success"
                    }
                }
            }
        },
        "/subjects/show/{vocab}/{id}": {
            "get": {
                "tags": [
                    "Authorities"
                ],
                "summary": "Find authority record by ID",
                "operationId": "App\\Http\\Controllers\\SubjectsController::show",
                "parameters": [
                    {
                        "name": "vocab",
                        "in": "path",
                        "description": "Subject vocabulary. Leave blank to search all subject vocabularies.",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "realfagstermer",
                                "humord",
                                "tekord",
                                "mrtermer",
                                "usvd",
                                "lskjema"
                            ]
                        }
                    },
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Local ID, e.g. `c006445`.",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "expand_mappings",
                        "in": "query",
                        "description": "If set to false, you'll only get the URIs for mapped concepts. If set to true, you will get data for one level of mappings.",
                        "schema": {
                            "type": "boolean",
                            "default": false
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success"
                    }
                }
            }
        },
        "/subjects/lookup": {
            "get": {
                "tags": [
                    "Authorities"
                ],
                "summary": "Find authority record by index term value",
                "description": "Get a single subject by term value.",
                "operationId": "App\\Http\\Controllers\\SubjectsController::lookup",
                "parameters": [
                    {
                        "name": "vocab",
                        "in": "query",
                        "description": "Subject vocabulary. Leave blank to search all subject vocabularies.",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "realfagstermer",
                                "humord",
                                "tekord",
                                "mrtermer",
                                "usvd",
                                "lskjema"
                            ]
                        }
                    },
                    {
                        "name": "query",
                        "in": "query",
                        "description": "Term, e.g. `Fisker`.",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "type",
                        "in": "query",
                        "description": "Example: Set type to ’Place’ if you want the place ‘Java’, ‘Topic’ if you want the programming language. ",
                        "schema": {
                            "type": "string",
                            "enum": [
                                "Concept",
                                "Facet",
                                "Topic",
                                "Place",
                                "Time",
                                "CompoundConcept",
                                "VirtualCompoundConcept",
                                "NonIndexable"
                            ]
                        }
                    },
                    {
                        "name": "expand_mappings",
                        "in": "query",
                        "description": "If set to false, you'll only get the URIs for mapped concepts. If set to true, you will get data for one level of mappings.",
                        "schema": {
                            "type": "boolean",
                            "default": false
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success"
                    }
                }
            }
        }
    },
...
}